requirements.txt
selenium==4.19.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
PyYAML==6.0.1
python-dotenv==1.0.1
openai==1.30.1
